requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.54.0
anyio==4.9.0
argcomplete==3.6.2
asyncpg==0.30.0
attrs==25.3.0
boto3==1.38.41
botocore==1.38.41
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
cohere==5.15.0
colorama==0.4.6
diskcache==5.6.3
distro==1.9.0
eval_type_backport==0.2.2
fasta2a==0.3.2
fastapi==0.115.13
fastavro==1.11.1
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
google-auth==2.40.3
google-generativeai
graphiti==0.1.13
graphiti-core==0.12.4
griffe==1.7.3
groq==0.28.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
jiter==0.10.0
jmespath==1.0.1
logfire-api==3.21.1
markdown-it-py==3.0.0
mcp==1.9.4
mdurl==0.1.2
mistralai==1.8.2
multidict==6.5.0
neo4j==5.28.1
numpy==2.3.1
openai==1.90.0
opentelemetry-api==1.34.1
packaging==25.0
pluggy==1.6.0
pandas>=2.0.0
plotly>=5.18.0

prompt_toolkit==3.0.51
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-ai==0.3.2
pydantic-ai-slim==0.3.2
pydantic-evals==0.3.2
pydantic-graph==0.3.2
pydantic-settings==2.10.0
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
pytest-asyncio==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
rsa==4.9.1
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
sse-starlette==2.3.6
starlette==0.46.2
streamlit>=1.30.0
tenacity
tiktoken
tokenizers==0.21.1
tqdm==4.67.1
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
wcwidth==0.2.13
websockets==15.0.1
yarl==1.20.1
zipp==3.23.0
